+++ R/data_interface.R
@@ -1,3 +1,24 @@
+# This function handles backward compatibility to support
+# unnamed datasets while not breaking sparklyr 0.9 param
+# signature. Returns a c(name, path) tuple.
+spark_read_compat_param <- function(sc, name, path) {
+  if (is.null(name) && is.null(path)) {
+    stop("The 'path' parameter must be specified.")
+  }
+  else if (identical(name, path)) {
+    # This is an invalid use case, for 'spark_read_*(cs, "hello")';
+    # however, for convenience and backwards compatibility we allow
+    # to use the second parameter as the path.
+    c(
+      spark_sanitize_names(tools::file_path_sans_ext(basename(name)), sc$config),
+      name
+    )
+  }
+  else {
+    c(name, path)
+  }
+}
+
 spark_csv_options <- function(header,
                               inferSchema,
                               delimiter,
@@ -66,8 +87,8 @@ spark_csv_options <- function(header,
 #'
 #' @export
 spark_read_csv <- function(sc,
-                           name,
-                           path,
+                           name = NULL,
+                           path = name,
                            header = TRUE,
                            columns = NULL,
                            infer_schema = TRUE,
@@ -81,6 +102,8 @@ spark_read_csv <- function(sc,
                            memory = TRUE,
                            overwrite = TRUE,
                            ...) {
+  c(name, path) %<-% spark_read_compat_param(sc, name, path)
+
   columnsHaveTypes <- length(names(columns)) > 0
   if (!identical(columns, NULL) & isTRUE(infer_schema) & columnsHaveTypes) {
     stop("'infer_schema' must be set to FALSE when 'columns' specifies column types")
@@ -201,8 +224,8 @@ spark_write_csv.spark_jobj <- function(x,
 #'
 #' @export
 spark_read_parquet <- function(sc,
-                               name,
-                               path,
+                               name = NULL,
+                               path = name,
                                options = list(),
                                repartition = 0,
                                memory = TRUE,
@@ -210,7 +233,7 @@ spark_read_parquet <- function(sc,
                                columns = NULL,
                                schema = NULL,
                                ...) {
-
+  c(name, path) %<-% spark_read_compat_param(sc, name, path)
   if (overwrite) spark_remove_table_if_exists(sc, name)
 
   df <- spark_data_read_generic(sc, list(spark_normalize_path(path)), "parquet", options, columns, schema)
@@ -284,15 +307,15 @@ spark_write_parquet.spark_jobj <- function(x,
 #'
 #' @export
 spark_read_json <- function(sc,
-                            name,
-                            path,
+                            name = NULL,
+                            path = name,
                             options = list(),
                             repartition = 0,
                             memory = TRUE,
                             overwrite = TRUE,
                             columns = NULL,
                             ...) {
-
+  c(name, path) %<-% spark_read_compat_param(sc, name, path)
   if (overwrite) spark_remove_table_if_exists(sc, name)
 
   df <- spark_data_read_generic(sc, spark_normalize_path(path), "json", options, columns)
@@ -591,13 +614,13 @@ spark_read_jdbc <- function(sc,
 #'
 #' @export
 spark_read_libsvm <- function(sc,
-                            name,
-                            path,
+                            name = NULL,
+                            path = name,
                             repartition = 0,
                             memory = TRUE,
                             overwrite = TRUE,
                             ...) {
-
+  c(name, path) %<-% spark_read_compat_param(sc, name, path)
   if (overwrite) spark_remove_table_if_exists(sc, name)
 
   df <- spark_data_read_generic(sc, "libsvm", "format") %>%
@@ -617,8 +640,8 @@ spark_read_libsvm <- function(sc,
 #'
 #' @export
 spark_read_source <- function(sc,
-                            name,
-                            path = NULL,
+                            name = NULL,
+                            path = name,
                             source,
                             options = list(),
                             repartition = 0,
@@ -626,7 +649,7 @@ spark_read_source <- function(sc,
                             overwrite = TRUE,
                             columns = NULL,
                             ...) {
-
+  c(name, path) %<-% spark_read_compat_param(sc, name, path)
   if (overwrite) spark_remove_table_if_exists(sc, name)
 
   df_reader <- spark_data_read_generic(sc, source, "format", options, columns)
@@ -745,15 +768,15 @@ spark_write_source.spark_jobj <- function(x,
 #'
 #' @export
 spark_read_text <- function(sc,
-                            name,
-                            path,
+                            name = NULL,
+                            path = name,
                             repartition = 0,
                             memory = TRUE,
                             overwrite = TRUE,
                             options = list(),
                             whole = FALSE,
                             ...) {
-
+  c(name, path) %<-% spark_read_compat_param(sc, name, path)
   if (overwrite) spark_remove_table_if_exists(sc, name)
 
   columns = list(line = "character")
@@ -830,8 +853,8 @@ spark_write_text.spark_jobj <- function(x,
 #'
 #' @export
 spark_read_orc <- function(sc,
-                           name,
-                           path,
+                           name = NULL,
+                           path = name,
                            options = list(),
                            repartition = 0,
                            memory = TRUE,
@@ -839,7 +862,7 @@ spark_read_orc <- function(sc,
                            columns = NULL,
                            schema = NULL,
                            ...) {
-
+  c(name, path) %<-% spark_read_compat_param(sc, name, path)
   if (overwrite) spark_remove_table_if_exists(sc, name)
 
   df <- spark_data_read_generic(sc, list(spark_normalize_path(path)), "orc", options, columns, schema)